refactor ui and controller module
