+++ src/controller/handlers.rs
@@ -1,0 +1,843 @@
+use super::*;
+
+impl Controller {
+    #[allow(clippy::missing_errors_doc, clippy::too_many_lines)]
+    pub fn handle_audio_event(
+        &mut self,
+        cx: &mut App,
+        event: &AudioEvent,
+        view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            AudioEvent::Position(pos) => {
+                let last_pos = self.state.read(cx).playback.position;
+
+                if *pos != last_pos {
+                    view.update(cx, |this, cx| {
+                        this.player_page.update(cx, |this, cx| {
+                            this.controlbar.update(cx, |this, cx| {
+                                this.playback_slider_state.update(cx, |this, cx| {
+                                    let state = cx.global::<Controller>().state.read(cx);
+                                    let current = if let Some(id) = state.playback.current {
+                                        state.library.tracks.get(&id)
+                                    } else {
+                                        None
+                                    };
+
+                                    let duration = if let Some(track) = current {
+                                        track.duration
+                                    } else {
+                                        Duration::default()
+                                    };
+                                    this.set_value(duration_to_slider(*pos, duration), cx);
+                                    cx.notify();
+                                });
+                            });
+                            // this.lyrics.update(cx, |this, cx| {
+                            //     let playing =
+                            //         self.state.read(cx).playback.status == PlaybackStatus::Playing;
+
+                            //     this.synchronize_playback(*pos, playing);
+
+                            //     if !playing {
+                            //         cx.notify();
+                            //     }
+                            // });
+                        });
+                        cx.notify();
+                    });
+                    self.state.update(cx, |this, cx| {
+                        this.playback.position = *pos;
+                        cx.notify();
+                    });
+
+                    self.system_integration_tx
+                        .send(SystemIntegrationCommand::SetPosition(*pos))
+                        .ok();
+
+                    let state = self.state.read(cx).playback.clone();
+                    let _ = self
+                        .cacher_tx
+                        .send(CacherCommand::WritePlaybackState(state));
+                }
+            }
+            AudioEvent::TrackLoaded(track_id, path) => {
+                let state = self.state.read(cx);
+                if !state.library.tracks.contains_key(track_id) {
+                    let _ = self
+                        .scanner_tx
+                        .send(ScannerCommand::ScanTrack(path.clone()));
+                }
+
+                if let Some(track) = state.library.tracks.get(track_id) {
+                    if let Some(image_id) = track.image_id {
+                        let _ = self.cacher_tx.send(CacherCommand::GetImage(
+                            HashSet::from([image_id]),
+                            ImageKind::AlbumArt,
+                        ));
+                    } else {
+                        let _ = self.image_processor_tx.send(
+                            ImageProcessorCommand::GetCurrentAlbumArt(*track_id, path.clone()),
+                        );
+                    }
+
+                    self.system_integration_tx
+                        .send(SystemIntegrationCommand::SetMetadata {
+                            title: track.title.clone(),
+                            artist: track.artist.clone(),
+                            album: track.album.clone(),
+                            image: None,
+                            duration: track.duration.as_secs(),
+                        })
+                        .ok();
+
+                    self.cacher_tx
+                        .send(CacherCommand::GetLyrics(*track_id))
+                        .ok();
+
+                    let lyrics_state = cx.global::<LyricsState>().0.clone();
+
+                    lyrics_state.update(cx, |this, cx| {
+                        this.status = LyricsStatus::Fetching;
+                        this.lyrics = None;
+                        this.track_id = Some(*track_id);
+
+                        cx.notify();
+                    });
+                }
+                self.state.update(cx, |this, cx| {
+                    this.playback.current = Some(*track_id);
+
+                    if let Some(idx) = this.queue.get_index(*track_id) {
+                        this.playback.current_index = idx;
+                    }
+
+                    cx.notify();
+                });
+
+                let state = self.state.read(cx).playback.clone();
+                let _ = self
+                    .cacher_tx
+                    .send(CacherCommand::WritePlaybackState(state));
+            }
+            AudioEvent::PlaybackStatus(status) => {
+                self.state.update(cx, |this, cx| {
+                    this.playback.status = *status;
+                    cx.notify();
+                });
+                cx.notify(view.entity_id());
+                let state = self.state.read(cx).playback.clone();
+                self.system_integration_tx
+                    .send(SystemIntegrationCommand::SetPlaybackStatus(
+                        *status,
+                        state.position,
+                    ))
+                    .ok();
+                let _ = self
+                    .cacher_tx
+                    .send(CacherCommand::WritePlaybackState(state));
+            }
+            AudioEvent::TrackEnded => {
+                let repeat = self.state.read(cx).playback.repeat;
+
+                if repeat {
+                    self.load_queue_current(cx);
+                } else {
+                    self.next(cx);
+                }
+            }
+        }
+        Ok(())
+    }
+
+    #[allow(clippy::missing_errors_doc, clippy::too_many_lines)]
+    pub fn handle_scanner_event(
+        &mut self,
+        cx: &mut App,
+        event: &ScannerEvent,
+        view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            ScannerEvent::UpsertTracks(tracks) => {
+                let mut modified_playlists = HashSet::new();
+                self.state.update(cx, |this, cx| {
+                    this.library.tracks.reserve(tracks.len());
+                    for (track, playlist_id) in tracks {
+                        let id = track.id;
+
+                        if let Some(existing) = this.library.tracks.get_mut(&id) {
+                            let existing = Arc::make_mut(existing);
+
+                            for src in &track.sources {
+                                if !existing.sources.iter().any(|s| s.path == src.path) {
+                                    existing.sources.push(src.clone());
+                                }
+                            }
+
+                            if existing.title.is_empty() && !track.title.is_empty() {
+                                existing.title.clone_from(&track.title);
+                            }
+
+                            if existing.artist.is_empty() && !track.artist.is_empty() {
+                                existing.artist.clone_from(&track.artist);
+                            }
+
+                            if existing.album.is_empty() && !track.album.is_empty() {
+                                existing.album.clone_from(&track.album);
+                            }
+                        } else {
+                            this.library.tracks.insert(id, Arc::new(track.clone()));
+                        }
+
+                        if let Some(pid) = playlist_id
+                            && let Some(playlist) = this.library.playlists.get_mut(pid)
+                        {
+                            if !playlist.tracks.contains(&id) {
+                                playlist.tracks.push(id);
+                            }
+                            modified_playlists.insert(*pid);
+                        }
+                    }
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ScannerEvent::InsertTracksIntoPlaylist(pid, tids) => {
+                self.state.update(cx, |this, cx| {
+                    if let Some(playlist) = this.library.playlists.get_mut(pid) {
+                        for tid in tids {
+                            if !playlist.tracks.contains(tid) {
+                                playlist.tracks.push(*tid);
+                            }
+                        }
+                    }
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ScannerEvent::AddTrackSource(id, source) => {
+                self.state.update(cx, |this, cx| {
+                    if let Some(track) = this.library.tracks.get_mut(id) {
+                        Arc::make_mut(track).sources.push(source.clone());
+                    }
+
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ScannerEvent::RemoveTrackSource(id, path) => {
+                self.state.update(cx, |this, cx| {
+                    if let Some(track) = this.library.tracks.get_mut(id)
+                        && let Some(source) =
+                            track.sources.iter().position(|this| this.path == *path)
+                    {
+                        Arc::make_mut(track).sources.remove(source);
+                    }
+
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ScannerEvent::InsertPlaylist(playlist) => {
+                self.state.update(cx, |this, cx| {
+                    this.library.playlists.insert(playlist.id, playlist.clone());
+
+                    cx.notify();
+                });
+
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ScannerEvent::ScanStarted => {
+                let scanning_status = cx.global_mut::<ScanningStatus>().clone().0;
+
+                scanning_status.update(cx, |this, cx| {
+                    this.is_scanning = true;
+                    this.is_discovering = true;
+
+                    cx.notify();
+                });
+
+                view.update(cx, |this, cx| {
+                    this.toast_manager.update(cx, |this, cx| {
+                        this.info("Scanning started...", cx);
+                        this.scanning_status(cx);
+                    });
+                    cx.notify();
+                });
+            }
+            ScannerEvent::Discovered(discovered) => {
+                let scanning_status = cx.global_mut::<ScanningStatus>().0.clone();
+
+                scanning_status.update(cx, |this, cx| {
+                    if !this.is_discovering {
+                        this.is_discovering = true;
+                    }
+
+                    this.discovered = *discovered;
+
+                    cx.notify();
+                });
+            }
+            ScannerEvent::Processed { processed, total } => {
+                let scanning_status = cx.global_mut::<ScanningStatus>().0.clone();
+
+                scanning_status.update(cx, |this, cx| {
+                    if this.is_discovering {
+                        this.is_discovering = false;
+                    }
+                    if !this.is_processing {
+                        this.is_processing = true;
+                    }
+
+                    this.total = *total;
+                    this.processed = *processed;
+                    cx.notify();
+                });
+            }
+            ScannerEvent::ScanFinished => {
+                self.scanner_tx.send(ScannerCommand::StartNextScan).ok();
+                let tracks = self.state.read(cx).library.tracks.clone();
+
+                let to_request: HashSet<(TrackId, PathBuf)> = tracks
+                    .iter()
+                    .filter(|(_, track)| track.image_id.is_none())
+                    .filter_map(|(id, track)| {
+                        track
+                            .get_valid_source()
+                            .map(|src| src.path.clone())
+                            .map(|path| (*id, path))
+                    })
+                    .collect();
+                let _ = self
+                    .image_processor_tx
+                    .send(ImageProcessorCommand::GetThumbnails(
+                        to_request,
+                        ImageKind::ThumbnailSmall,
+                    ));
+
+                view.update(cx, |this, cx| {
+                    this.toast_manager.update(cx, |this, cx| {
+                        this.toasts.update(cx, |list, _| {
+                            for t in list.iter_mut() {
+                                if matches!(t.kind, ToastKind::ScanProgress(_))
+                                    && t.phase != ToastPhase::Exiting
+                                {
+                                    t.phase = ToastPhase::Exiting;
+                                    t.exiting_at = Some(Instant::now());
+                                }
+                            }
+                        });
+                        this.success("Scan complete!", cx);
+                    });
+                });
+
+                let status = cx.global::<ScanningStatus>().0.clone();
+                status.update(cx, |s, _| {
+                    s.is_scanning = false;
+                    s.is_discovering = false;
+                    s.is_processing = false;
+
+                    s.discovered = 0;
+                    s.total = 0;
+                    s.processed = 0;
+                });
+
+                let library = cx.global::<Controller>().state.read(cx).library.clone();
+                let missing: Vec<PlaylistId> = library
+                    .playlists
+                    .iter()
+                    .filter_map(|(id, playlist)| {
+                        if playlist.image_id.is_none() {
+                            Some(*id)
+                        } else {
+                            None
+                        }
+                    })
+                    .collect();
+                self.request_playlist_thumbnails(&missing, cx);
+            }
+        }
+        Ok(())
+    }
+
+    #[allow(clippy::missing_errors_doc)]
+    pub fn handle_image_processor_event(
+        &mut self,
+        cx: &mut App,
+        event: &ImageProcessorEvent,
+        view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            ImageProcessorEvent::InsertAlbumArt(image_id, image) => {
+                let image_cache = cx.global_mut::<ImageCache>();
+                image_cache.current = Some(image.clone());
+
+                let image = image.clone();
+                let width = image.size(0).width.0.cast_unsigned();
+                let height = image.size(0).height.0.cast_unsigned();
+                if let Some(image) = image.as_bytes(0) {
+                    let image = image.to_vec();
+                    let _ = self.cacher_tx.send(CacherCommand::WriteImage {
+                        id: *image_id,
+                        kind: ImageKind::AlbumArt,
+                        width,
+                        height,
+                        image: image.clone(),
+                    });
+                    let state = self.state.read(cx);
+                    if let Some(track_id) = &state.playback.current
+                        && let Some(track) = state.library.tracks.get(track_id)
+                    {
+                        self.system_integration_tx
+                            .send(SystemIntegrationCommand::SetMetadata {
+                                title: track.title.clone(),
+                                artist: track.artist.clone(),
+                                album: track.album.clone(),
+                                image: Some((width, height, image)),
+                                duration: track.duration.as_secs(),
+                            })
+                            .ok();
+                    }
+                }
+
+                cx.notify(view.entity_id());
+            }
+            ImageProcessorEvent::InsertThumbnails(thumbnails, kind) => {
+                for (id, image) in thumbnails {
+                    let width = image.size(0).width.0.cast_unsigned();
+                    let height = image.size(0).height.0.cast_unsigned();
+                    if let Some(image) = image.as_bytes(0) {
+                        let image = image.to_vec();
+                        let _ = self.cacher_tx.send(CacherCommand::WriteImage {
+                            id: *id,
+                            kind: *kind,
+                            width,
+                            height,
+                            image,
+                        });
+                    }
+
+                    let evicted = {
+                        let thumbnail_cache = cx.global_mut::<ImageCache>();
+                        thumbnail_cache.inflight.remove(id);
+                        thumbnail_cache.add(*id, image.clone())
+                    };
+
+                    if let Some(img) = evicted {
+                        drop_image_from_app(cx, img);
+                    }
+                }
+                cx.notify(view.entity_id());
+            }
+            ImageProcessorEvent::UpdateImageLookup(lookup) => {
+                self.state.update(cx, |this, cx| {
+                    for (id, image_id) in lookup {
+                        if let Some(track) = this.library.tracks.get_mut(id) {
+                            Arc::make_mut(track).image_id = Some(*image_id);
+                        }
+                    }
+
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+            ImageProcessorEvent::InsertPlaylistThumbnail(id, image_id, image) => {
+                let thumbnail_cache = cx.global_mut::<ImageCache>();
+
+                thumbnail_cache.add(*image_id, image.clone());
+
+                thumbnail_cache.inflight.remove(image_id);
+
+                let _ = self
+                    .image_processor_tx
+                    .send(ImageProcessorCommand::PlaylistJobFinished(*id));
+
+                let width = image.size(0).width.0.cast_unsigned();
+                let height = image.size(0).height.0.cast_unsigned();
+                if let Some(image) = image.as_bytes(0) {
+                    let image = image.to_vec();
+                    let _ = self.cacher_tx.send(CacherCommand::WriteImage {
+                        id: *image_id,
+                        kind: ImageKind::Playlist,
+                        width,
+                        height,
+                        image,
+                    });
+                }
+
+                self.state.update(cx, |this, cx| {
+                    if let Some(playlist) = this.library.playlists.get_mut(id) {
+                        playlist.image_id = Some(*image_id);
+                    }
+                    cx.notify();
+                });
+                let state = self.state.read(cx).library.clone();
+                let _ = self.cacher_tx.send(CacherCommand::WriteLibraryState(state));
+            }
+        }
+
+        Ok(())
+    }
+
+    #[allow(clippy::missing_errors_doc, clippy::too_many_lines)]
+    pub fn handle_cacher_event(
+        &mut self,
+        cx: &mut App,
+        event: &CacherEvent,
+        view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            CacherEvent::AppState(state) => {
+                let playback_state = state.playback.clone();
+                self.state.update(cx, |this, _| {
+                    *this = state.clone();
+                });
+
+                self.load_queue_current(cx);
+                self.set_volume(playback_state.volume, cx);
+                self.seek(playback_state.position);
+
+                match playback_state.status {
+                    PlaybackStatus::Stopped => self.stop(),
+                    PlaybackStatus::Paused => self.pause(),
+                    PlaybackStatus::Playing => self.play(),
+                }
+
+                let duration = if let Some(current) = playback_state.current
+                    && let Some(track) = state.library.tracks.get(&current)
+                {
+                    Some(track.duration)
+                } else {
+                    None
+                };
+
+                view.update(cx, |this, cx| {
+                    this.player_page.update(cx, |this, cx| {
+                        this.controlbar.update(cx, |this, cx| {
+                            this.vol_slider_state.update(cx, |this, cx| {
+                                this.set_value(playback_state.volume * 100.0, cx);
+                            });
+                            this.playback_slider_state.update(cx, |this, cx| {
+                                if let Some(duration) = duration {
+                                    this.set_value(
+                                        duration_to_slider(playback_state.position, duration),
+                                        cx,
+                                    );
+                                }
+                            });
+                        });
+                    });
+                });
+            }
+            CacherEvent::Thumbnails(thumbnails) => {
+                for (id, image) in thumbnails {
+                    let evicted = {
+                        let thumbnail_cache = cx.global_mut::<ImageCache>();
+                        thumbnail_cache.inflight.remove(id);
+                        thumbnail_cache.add(*id, image.clone())
+                    };
+
+                    if let Some(img) = evicted {
+                        drop_image_from_app(cx, img);
+                    }
+                }
+                cx.notify(view.entity_id());
+            }
+            CacherEvent::AlbumArt(image) => {
+                let image_cache = cx.global_mut::<ImageCache>();
+
+                image_cache.current = Some(image.clone());
+
+                let image = image.clone();
+                let width = image.size(0).width.0.cast_unsigned();
+                let height = image.size(0).height.0.cast_unsigned();
+                if let Some(image) = image.as_bytes(0) {
+                    fn rgb_to_rgba(color: Rgb<u8>) -> Rgba {
+                        rgb((u32::from(color.r) << 16)
+                            | (u32::from(color.g) << 8)
+                            | u32::from(color.b))
+                    }
+
+                    let image = image.to_vec();
+                    let state = self.state.read(cx);
+                    if let Some(track_id) = &state.playback.current
+                        && let Some(track) = state.library.tracks.get(track_id)
+                    {
+                        self.system_integration_tx
+                            .send(SystemIntegrationCommand::SetMetadata {
+                                title: track.title.clone(),
+                                artist: track.artist.clone(),
+                                album: track.album.clone(),
+                                image: Some((width, height, image.clone())),
+                                duration: track.duration.as_secs(),
+                            })
+                            .ok();
+                    }
+
+                    let mut rgb_bytes = Vec::with_capacity((width * height * 3) as usize);
+                    for chunk in image.as_slice().chunks_exact(4) {
+                        rgb_bytes.push(chunk[0]);
+                        rgb_bytes.push(chunk[1]);
+                        rgb_bytes.push(chunk[2]);
+                    }
+
+                    let input =
+                        okmain::InputImage::from_bytes(width as u16, height as u16, &rgb_bytes)
+                            .unwrap();
+
+                    let colors = okmain::colors(input);
+
+                    let dominant = DominantColors {
+                        color1: colors.first().copied().map_or(rgb(0x000000), rgb_to_rgba),
+
+                        color2: colors.get(1).copied().map_or(rgb(0x000000), rgb_to_rgba),
+
+                        color3: colors.get(2).copied().map_or(rgb(0x000000), rgb_to_rgba),
+
+                        color4: colors.get(3).copied().map_or(rgb(0x000000), rgb_to_rgba),
+                    };
+                    *cx.global_mut::<DominantColors>() = dominant;
+                }
+                cx.notify(view.entity_id());
+            }
+            CacherEvent::PlaylistThumbnail(id, thumbnail) => {
+                cx.global_mut::<ImageCache>().inflight.remove(id);
+
+                let evicted = {
+                    let image_cache = cx.global_mut::<ImageCache>();
+                    image_cache.add(*id, thumbnail.clone())
+                };
+
+                if let Some(img) = evicted {
+                    drop_image_from_app(cx, img);
+                }
+                cx.notify(view.entity_id());
+            }
+            CacherEvent::MissingAlbumArt(id) => {
+                cx.global_mut::<ImageCache>().inflight.remove(id);
+
+                let state = self.state.read(cx);
+                let tracks = state.library.tracks.clone();
+
+                let track_id = tracks.iter().find_map(|(tid, track)| {
+                    if track.image_id == Some(*id) {
+                        Some(tid)
+                    } else {
+                        None
+                    }
+                });
+
+                if let Some(track_id) = track_id
+                    && let Some(track) = tracks.get(track_id)
+                    && let Some(source) = track.get_valid_source()
+                {
+                    let _ =
+                        self.image_processor_tx
+                            .send(ImageProcessorCommand::GetCurrentAlbumArt(
+                                *track_id,
+                                source.path.clone(),
+                            ));
+                }
+            }
+            CacherEvent::MissingThumbnails(ids) => {
+                let cache = cx.global_mut::<ImageCache>();
+
+                for id in ids {
+                    cache.inflight.remove(id);
+                }
+
+                let state = self.state.read(cx);
+                let tracks = state.library.tracks.clone();
+
+                for id in ids {
+                    let track_id = tracks.iter().find_map(|(tid, track)| {
+                        if track.image_id == Some(*id) {
+                            Some(tid)
+                        } else {
+                            None
+                        }
+                    });
+
+                    if let Some(track_id) = track_id
+                        && let Some(track) = tracks.get(track_id)
+                        && let Some(source) = track.get_valid_source()
+                    {
+                        let mut set = HashSet::new();
+                        set.insert((*track_id, source.path.clone()));
+                        let _ = self
+                            .image_processor_tx
+                            .send(ImageProcessorCommand::GetThumbnails(
+                                set,
+                                ImageKind::ThumbnailSmall,
+                            ));
+                    }
+                }
+            }
+            CacherEvent::MissingPlaylistThumbnail(id) => {
+                cx.global_mut::<ImageCache>().inflight.remove(id);
+
+                let state = self.state.read(cx);
+                let playlists = state.library.playlists.clone();
+
+                let playlist_id = playlists.iter().find_map(|(pid, playlist)| {
+                    if playlist.image_id == Some(*id) {
+                        Some(pid)
+                    } else {
+                        None
+                    }
+                });
+
+                if let Some(playlist_id) = playlist_id
+                    && let Some(playlist) = playlists.get(playlist_id)
+                {
+                    let playlist_tracks = playlist.tracks.clone();
+                    let thumb_tracks = {
+                        let state = self.state.read(cx);
+
+                        pick_playlist_thumbnail_tracks(&state.library.tracks, &playlist_tracks, 4)
+                    };
+
+                    let _ =
+                        self.image_processor_tx
+                            .send(ImageProcessorCommand::PlaylistThumbnail {
+                                id: *playlist_id,
+                                tracks: thumb_tracks,
+                            });
+                }
+            }
+            CacherEvent::Lyrics(id, lyrics) => {
+                let current = cx.global::<Controller>().state.read(cx).playback.current;
+
+                if let Some(current) = current
+                    && current == *id
+                {
+                    let lyrics_state = cx.global::<LyricsState>().0.clone();
+
+                    lyrics_state.update(cx, |this, cx| {
+                        this.lyrics.clone_from(lyrics);
+                        this.track_id = Some(current);
+                        this.status = if lyrics.is_some() {
+                            LyricsStatus::Available
+                        } else {
+                            LyricsStatus::Unavailable
+                        };
+                        cx.notify();
+                    });
+                }
+            }
+            CacherEvent::MissingLyrics(id) => {
+                if let Some(track) = self.state.read(cx).library.tracks.get(id) {
+                    self.get_lyrics(
+                        *id,
+                        &track.title,
+                        &track.artist,
+                        &track.album,
+                        track.duration,
+                    );
+                }
+            }
+        }
+        Ok(())
+    }
+
+    #[allow(clippy::missing_errors_doc, clippy::too_many_lines)]
+    pub fn handle_system_integration_event(
+        &mut self,
+        cx: &mut App,
+        event: &SystemIntegrationEvent,
+        _view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            SystemIntegrationEvent::PlayPause => {
+                let status = self.state.read(cx).playback.status;
+
+                if status == PlaybackStatus::Stopped || status == PlaybackStatus::Paused {
+                    self.play();
+                } else {
+                    self.pause();
+                }
+            }
+            SystemIntegrationEvent::Play => {
+                self.play();
+            }
+            SystemIntegrationEvent::Pause => {
+                self.pause();
+            }
+            SystemIntegrationEvent::Stop => {
+                self.stop();
+            }
+            SystemIntegrationEvent::Next => {
+                self.next(cx);
+            }
+            SystemIntegrationEvent::Prev => {
+                self.prev(cx);
+            }
+            SystemIntegrationEvent::SeekForward(duration) => {
+                let pos = self.state.read(cx).playback.position;
+
+                self.seek(pos.saturating_add(*duration));
+            }
+            SystemIntegrationEvent::SeekBackward(duration) => {
+                let pos = self.state.read(cx).playback.position;
+
+                self.seek(pos.saturating_sub(*duration));
+            }
+            SystemIntegrationEvent::Volume(vol) => {
+                #[allow(clippy::cast_possible_truncation)]
+                self.set_volume(*vol as f32, cx);
+            }
+            SystemIntegrationEvent::Position(pos) => {
+                self.seek(*pos);
+            }
+        }
+
+        Ok(())
+    }
+
+    #[allow(clippy::missing_errors_doc, clippy::too_many_lines)]
+    pub fn handle_lyrics_event(
+        &mut self,
+        cx: &mut App,
+        event: &LyricsEvent,
+        _view: &Entity<Wiremann>,
+    ) -> Result<(), ControllerError> {
+        match event {
+            LyricsEvent::Lyrics(id, lyrics) => {
+                let current = cx.global::<Controller>().state.read(cx).playback.current;
+
+                if let Some(current) = current
+                    && current == *id
+                {
+                    let lyrics_state = cx.global::<LyricsState>().0.clone();
+
+                    lyrics_state.update(cx, |this, cx| {
+                        this.lyrics.clone_from(lyrics);
+                        this.track_id = Some(current);
+
+                        this.status = if lyrics.is_some() {
+                            LyricsStatus::Available
+                        } else {
+                            LyricsStatus::Unavailable
+                        };
+
+                        cx.notify();
+                    });
+                }
+                if let Some(lyrics) = lyrics {
+                    self.cacher_tx
+                        .send(CacherCommand::WriteLyrics(*id, lyrics.clone()))
+                        .ok();
+                }
+            }
+        }
+
+        Ok(())
+    }
+}
+
