+++ src/ui/components/pages/mod.rs
@@ -1,3 +1,5 @@
 pub mod library;
+pub mod library_helpers;
 pub mod player;
 pub mod playlists;
+pub mod playlists_helpers;
